+++ NEWS.md
@@ -4,8 +4,7 @@
 
 - The exchange algorithm for optimizing diversity and variance
  objectives has been reimplemented in C, leading to a substantial 
- boost in performance. **Known problem: Does not seem to work
- under windows.**
+ boost in performance.
 
 # anticlust 0.5.2